Rolling back main to a safer time.
