+++ backend/audit/views.py
@@ -17,8 +17,9 @@ from config.settings import (
     SP_FRAMEWORK_BASIS,
     SP_FRAMEWORK_OPINIONS,
 )
-from audit.fixtures.excel import FORM_SECTIONS, UNKNOWN_WORKBOOK
+from .fixtures.excel import FORM_SECTIONS, UNKNOWN_WORKBOOK
 
+from audit.cross_validation import sac_validation_shape, submission_progress_check
 from audit.excel import (
     extract_additional_ueis,
     extract_additional_eins,
@@ -62,11 +63,7 @@ from audit.validators import (
     validate_notes_to_sefa_json,
     validate_secondary_auditors_json,
 )
-from audit.viewlib import (  # noqa
-    SubmissionProgressView,
-    UploadReportView,
-    submission_progress_check,
-)
+from audit.viewlib import UploadReportView  # noqa
 
 
 logging.basicConfig(
@@ -695,6 +692,85 @@ class SubmissionView(CertifyingAuditeeRequiredMixin, generic.View):
 
         except SingleAuditChecklist.DoesNotExist:
             raise PermissionDenied("You do not have access to this audit.")
+
+
+class SubmissionProgressView(SingleAuditChecklistAccessRequiredMixin, generic.View):
+    """
+    Display information about and the current status of the sections of the submission,
+    including links to the pages for the sections.
+
+    The following sections have three states, rather than two:
+
+    +   Additionai UEIs
+    +   Additionai EINs
+    +   Secondary Auditors
+
+    The states are:
+
+    +   hidden
+    +   incomplete
+    +   complete
+
+    In each case, they are hidden if the corresponding question in the General
+    Information form has been answered with a negative response.
+    """
+
+    def get(self, request, *args, **kwargs):
+        report_id = kwargs["report_id"]
+
+        try:
+            sac = SingleAuditChecklist.objects.get(report_id=report_id)
+            try:
+                sar = SingleAuditReportFile.objects.filter(sac_id=sac.id).latest(
+                    "date_created"
+                )
+            except SingleAuditReportFile.DoesNotExist:
+                sar = None
+
+            shaped_sac = sac_validation_shape(sac)
+            subcheck = submission_progress_check(shaped_sac, sar, crossval=False)
+
+            context = {
+                "single_audit_checklist": {
+                    "created": True,
+                    "created_date": sac.date_created,
+                    "created_by": sac.submitted_by,
+                    "completed": False,
+                    "completed_date": None,
+                    "completed_by": None,
+                },
+                "pre_submission_validation": {
+                    "completed": sac.submission_status == "ready_for_certification",
+                    "completed_date": None,
+                    "completed_by": None,
+                    # We want the user to always be able to run this check:
+                    "enabled": True,
+                },
+                "certification": {
+                    "auditor_certified": bool(sac.auditor_certification),
+                    "auditor_enabled": sac.submission_status
+                    == "ready_for_certification",
+                    "auditee_certified": bool(sac.auditee_certification),
+                    "auditee_enabled": sac.submission_status == "auditor_certified",
+                },
+                "submission": {
+                    "completed": sac.submission_status == "submitted",
+                    "completed_date": None,
+                    "completed_by": None,
+                    "enabled": sac.submission_status == "auditee_certified",
+                },
+                "report_id": report_id,
+                "auditee_name": sac.auditee_name,
+                "auditee_uei": sac.auditee_uei,
+                "user_provided_organization_type": sac.user_provided_organization_type,
+            }
+            context = context | subcheck
+
+            return render(
+                request, "audit/submission_checklist/submission-checklist.html", context
+            )
+        except SingleAuditChecklist.DoesNotExist as err:
+            raise PermissionDenied("You do not have access to this audit.") from err
 
 
 class AuditInfoFormView(SingleAuditChecklistAccessRequiredMixin, generic.View):
